UPDATED: Renamed files, such that the naming convention is the same everywhere.
