+++ src/gui/Sidebar.tsx
