+++ src/gui/StatsModal.tsx
